Finish Auth

+++ Helpers/JWT.cs
@@ -5,7 +5,7 @@
 		public string Key { get; set; }
 		public string Issuer { get; set; }
 		public string Audience { get; set; }
-		public int DurationInDays { get; set; }
+		public double DurationInDays { get; set; }
 		
 
 	}
